Fix last_seen

+++ types/tgl_user.h
@@ -42,7 +42,7 @@ enum class tgl_user_online_status: int32_t {
 
 struct tgl_user_status {
     tgl_user_online_status online;
-    int32_t when;
+    int64_t when;
     tgl_user_status(): online(tgl_user_online_status::unknown), when(0) { }
 };
